I know this is not what you're supposed to do for database migrations but I don't have time to dive into Alembic setup
